Improve ImPlot performance by using the newly available API.

+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/gdx/ui/yo/ImPlotPlotLineSwapBuffer.java
@@ -14,5 +14,5 @@ public interface ImPlotPlotLineSwapBuffer
 
    public void copyPreviousToUpdated(int srcPos, int destPos, int length);
 
-   public void plot(String labelID, Integer[] xValues, int offset);
+   public void plot(String labelID, double[] xValues, int offset);
 }
